doi db qua mysql

+++ ViewModel/HomeVM.cs
@@ -6,8 +6,9 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Img { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string Description { get; set; }
         public decimal KhoiLuong { get; set; }
     }
+
 }
